move offline imagery-related files to core:domain
