+++ app/src/main/java/org/groundplatform/android/ui/map/MapFragment.kt
@@ -18,7 +18,7 @@ package org.groundplatform.android.ui.map
 import android.annotation.SuppressLint
 import androidx.annotation.IdRes
 import kotlinx.coroutines.flow.SharedFlow
-import org.groundplatform.android.model.imagery.TileSource
+import org.groundplatform.domain.model.imagery.TileSource
 import org.groundplatform.android.ui.common.AbstractFragment
 import org.groundplatform.domain.model.geometry.Coordinates
 import org.groundplatform.domain.model.map.Bounds
