+++ app/src/main/java/org/groundplatform/android/ui/map/gms/ClippingTileProvider.kt
@@ -21,9 +21,9 @@ import com.google.android.gms.maps.model.Tile
 import com.google.android.gms.maps.model.TileProvider
 import com.google.android.gms.maps.model.TileProvider.NO_TILE
 import org.groundplatform.android.ui.map.gms.mog.ImageEditor
-import org.groundplatform.android.ui.map.gms.mog.TileCoordinates
 import org.groundplatform.android.ui.map.gms.mog.toPixelBounds
-import org.groundplatform.android.ui.map.gms.mog.toPixelCoordinate
+import org.groundplatform.domain.model.imagery.TileCoordinates
+import org.groundplatform.domain.model.imagery.toPixelCoordinate
 
 private const val MAX_ZOOM = 19
