+++ app/src/main/java/org/groundplatform/android/ui/map/gms/mog/MogTileDownloader.kt
@@ -21,6 +21,7 @@ import kotlinx.coroutines.currentCoroutineContext
 import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.cancellable
 import kotlinx.coroutines.flow.flow
+import org.groundplatform.domain.model.imagery.TileCoordinates
 import timber.log.Timber
 
 /**
