+++ app/src/main/java/org/groundplatform/android/ui/offlineareas/OfflineAreasViewModel.kt
@@ -24,11 +24,11 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.launch
-import org.groundplatform.android.model.imagery.OfflineArea
+import org.groundplatform.domain.model.imagery.OfflineArea
 import org.groundplatform.android.repository.OfflineAreaRepository
 import org.groundplatform.android.ui.common.AbstractViewModel
-import org.groundplatform.android.util.toMb
-import org.groundplatform.android.util.toMbString
+import org.groundplatform.android.ui.util.toMbString
+import org.groundplatform.domain.model.util.toMb
 
 /**
  * View model for the offline area manager fragment. Handles the current list of downloaded areas.
