+++ app/src/main/java/org/groundplatform/android/ui/offlineareas/selector/OfflineAreaSelectorViewModel.kt
@@ -26,8 +26,8 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.launch
 import org.groundplatform.android.di.coroutines.IoDispatcher
-import org.groundplatform.android.model.imagery.RemoteMogTileSource
-import org.groundplatform.android.model.imagery.TileSource
+import org.groundplatform.domain.model.imagery.RemoteMogTileSource
+import org.groundplatform.domain.model.imagery.TileSource
 import org.groundplatform.android.repository.OfflineAreaRepository
 import org.groundplatform.android.system.LocationManager
 import org.groundplatform.android.system.NetworkManager
@@ -36,10 +36,10 @@ import org.groundplatform.android.system.SettingsManager
 import org.groundplatform.android.ui.common.BaseMapViewModel
 import org.groundplatform.android.ui.offlineareas.selector.model.OfflineAreaSelectorEvent
 import org.groundplatform.android.ui.offlineareas.selector.model.OfflineAreaSelectorState
-import org.groundplatform.android.util.toMb
-import org.groundplatform.android.util.toMbString
+import org.groundplatform.android.ui.util.toMbString
 import org.groundplatform.domain.model.map.Bounds
 import org.groundplatform.domain.model.map.CameraPosition
+import org.groundplatform.domain.model.util.toMb
 import org.groundplatform.domain.repository.LocationOfInterestRepositoryInterface
 import org.groundplatform.domain.repository.MapStateRepositoryInterface
 import org.groundplatform.domain.repository.SurveyRepositoryInterface
