+++ app/src/test/java/org/groundplatform/android/data/local/LocalOfflineAreaStore.kt
@@ -21,7 +21,7 @@ import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.data.local.stores.LocalOfflineAreaStore
-import org.groundplatform.android.model.imagery.OfflineArea
+import org.groundplatform.domain.model.imagery.OfflineArea
 import org.groundplatform.domain.model.map.Bounds
 import org.junit.Test
 import org.junit.runner.RunWith
