+++ app/src/test/java/org/groundplatform/android/ui/map/gms/mog/MogClientTest.kt
@@ -24,6 +24,7 @@ import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.runBlocking
 import org.groundplatform.android.data.remote.RemoteStorageManager
 import org.groundplatform.domain.model.geometry.Coordinates
+import org.groundplatform.domain.model.imagery.TileCoordinates
 import org.groundplatform.domain.model.map.Bounds
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
