+++ app/src/test/java/org/groundplatform/android/ui/map/gms/mog/MogImageMetadataTest.kt
@@ -16,6 +16,7 @@
 package org.groundplatform.android.ui.map.gms.mog
 
 import com.google.common.truth.Truth.assertThat
+import org.groundplatform.domain.model.imagery.TileCoordinates
 import org.junit.Assert.assertThrows
 import org.junit.Test
