+++ app/src/test/java/org/groundplatform/android/ui/map/gms/mog/PixelCoordinatesTest.kt
@@ -15,8 +15,12 @@
  */
 package org.groundplatform.android.ui.map.gms.mog
 
-import com.google.android.gms.maps.model.LatLng
 import kotlin.test.assertEquals
+import org.groundplatform.domain.model.geometry.Coordinates
+import org.groundplatform.domain.model.imagery.PixelCoordinates
+import org.groundplatform.domain.model.imagery.TileCoordinates
+import org.groundplatform.domain.model.imagery.toPixelCoordinate
+import org.groundplatform.domain.model.imagery.toPixelCoordinates
 import org.junit.Test
 
 class PixelCoordinatesTest {
@@ -44,7 +48,7 @@ class PixelCoordinatesTest {
   @Test
   fun `LatLng#toPixelCoordinates() at zoom 0 returns correct values`() {
     // jsfiddle used to find examples: https://jsfiddle.net/gmiceli/pjhy4Lfm/16/
-    val actual = LatLng(40.6874, -73.9306).toPixelCoordinates(0)
+    val actual = Coordinates(40.6874, -73.9306).toPixelCoordinates(0)
     val expected = PixelCoordinates(75, 96, 0)
     assertEquals(expected, actual)
   }
@@ -52,7 +56,7 @@ class PixelCoordinatesTest {
   @Test
   fun `LatLng#toPixelCoordinates() at intermediate zoom returns correct values`() {
     // jsfiddle used to find examples: https://jsfiddle.net/gmiceli/pjhy4Lfm/16/
-    val actual = LatLng(41.876, 12.4757).toPixelCoordinates(3)
+    val actual = Coordinates(41.876, 12.4757).toPixelCoordinates(3)
     val expected = PixelCoordinates(1094, 761, 3)
     assertEquals(expected, actual)
   }
