+++ app/src/test/java/org/groundplatform/android/ui/map/gms/mog/TiffTagDataTypeTest.kt
@@ -16,6 +16,7 @@
 package org.groundplatform.android.ui.map.gms.mog
 
 import com.google.common.truth.Truth.assertThat
+import org.groundplatform.domain.model.imagery.TiffTagDataType
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
