+++ core/domain/src/commonMain/kotlin/org/groundplatform/domain/model/imagery/TileSource.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.android.model.imagery
+package org.groundplatform.domain.model.imagery
 
 import org.groundplatform.domain.model.map.Bounds
