+++ core/domain/src/commonMain/kotlin/org/groundplatform/domain/model/util/DataSize.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Google LLC
+ * Copyright 2026 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.android.ui.map.gms.mog
+package org.groundplatform.domain.model.util
 
-import kotlin.math.abs
-import kotlin.math.cos
+typealias ByteCount = Int
 
-/**
- * Returns the value shifted left [n] bits when [n] is positive, or right [-n] bits when negative.
- */
-fun Int.shiftLeft(n: Int) = if (n >= 0) this shl n else this shr abs(n)
+typealias MegabyteCount = Float
 
-/** Returns the secant of angle `x` given in radians. */
-fun sec(x: Double) = 1 / cos(x)
-
-/** Converts degrees into radians. */
-fun Double.toRadians() = this * (Math.PI / 180)
+/** Returns the byte count as an equivalent megabyte count. */
+fun ByteCount.toMb(): MegabyteCount = this / (1024f * 1024f)
